fix: check if .devkit/contracts is inline (#66)
Currently, we're checking to see if `.devkit/contracts` is a submodule
by checking if the `.git/modules/...` directory is present, if an
upgrade is attempted and fails and the subsequent fix commit is removed
to get back to a clean slate, then the `.git/modules/...` directory
persists and the check fails essentially bricking the project.

This PR uses git directly to check if the `.devkit/contracts` directory
is registered as a submodule or not, to prevent any reliance on
faulty/flaky state.